Move SettingConfig and SettingGroup types to src/types/settings.ts
